requirements.txt
Django==4.2
django-mptt==0.14.0